scripts/requirements.txt
# Python requirements for DentalAI ML scripts
# Install with: pip install -r requirements.txt

# Core ML dependencies
torch>=2.0.0
coremltools>=7.0
ultralytics>=8.0.0

# Data processing
numpy>=1.21.0
opencv-python>=4.5.0
pillow>=8.0.0

# Configuration
pyyaml>=6.0

# Optional: For advanced features
# tensorflow>=2.10.0  # Alternative to PyTorch
# onnx>=1.12.0       # ONNX format support
# matplotlib>=3.5.0  # Plotting and visualization
# seaborn>=0.11.0    # Statistical visualization
